refactor(slack): consolidate message posting logic into a dedicated utility
- Replaced the `slackPostMessageHandler` with a new `postSlackMessage` utility to streamline message posting functionality.
- Updated references in `postSlackChannelWelcome` to utilize the new utility, enhancing code clarity and maintainability.
- Removed redundant imports and improved the structure of the message handling process.
